Junk ScriptableObject

+++ Assets/_Data/Damage/DamageSender.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class DamageSender : AnMonoBehaviour
 {
-    [SerializeField] protected float damage = 1;
+    [SerializeField] protected int damage = 1;
     public virtual void Send(Transform obj)
     {
         DamageReceiver damageReceiver = obj.GetComponentInChildren<DamageReceiver>();
